[#20250104] Move domains routes to dashboard
* Relocates domain-related routes from account to dashboard routes
* Updates navigation link from '/account/domains' to '/domains'
* Renames AccountDomains component to DashboardDomains
* Maintains existing functionality and auth requirements
